Silence a false-positive from Coverity Scan
Coverity Scan would emit a warning when run on a platform with
CLOCK_GETTIME because, on that code path, there was an unreachable
return statement.  The statement gets hit on other platforms, so it is
not an issue, but we don't want to cause noise for people including
tinycthread in their code.