Use _ftime_s instead of _ftime64_s
struct _timeb is not equivalent to struct _timeb64 on mingw, passing it
to _ftime64_s results in an error.  However, on real Windows, _ftime_s
should be mapped to _ftime64_s (unless time_t is 32-bits, in which case
we want a 32-bit number anyway), so AFAICT there is no real benefit to
calling _ftime64_s.