Fix deadlock in test_mutex_timed
When building tests in the Release build type, the call to thrd_sleep
would get removed when contained in an assert macro. The lack of
sleep would lead to a deadlock. Move the call to thrd_sleep to be
outside the macro, so that the mutex-timed test can still complete
in a Release build.